java/src/test/java/com/yourtion/leetcode/daily/m10/d30/SolutionTest.java
package com.yourtion.leetcode.daily.m10.d30;

import com.yourtion.leetcode.TestUtils;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20201030")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[[0,1,0,0],[1,1,1,0],[0,1,0,0],[1,1,0,0]]", 16)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void islandPerimeter(String source, int res) {
        System.out.printf("runTest: %s , res: %d", source, res);
        int[][] ss = TestUtils.stringToInt2dArray(source);
        Assertions.assertEquals(res, new Solution().islandPerimeter(ss));
    }
}